+++ src/main/java/com/mposhatov/entity/MainCharacteristics.java
@@ -6,71 +6,71 @@ import javax.persistence.*;
 public class MainCharacteristics {
 
     @Column(name = "ATTACK", nullable = false)
-    protected long attack;
+    protected int attack;
 
     @Column(name = "PHYSICAL_DEFENSE", nullable = false)
-    protected long physicalDefense;
+    protected int physicalDefense;
 
     @Column(name = "MAGIC_DEFENSE", nullable = false)
-    protected long magicDefense;
+    protected int magicDefense;
 
     @Column(name = "SPELL_POWER", nullable = false)
-    protected long spellPower;
+    protected int spellPower;
 
     @Column(name = "MANA", nullable = false)
-    protected long mana;
+    protected int mana;
 
     protected MainCharacteristics() {
     }
 
-    public MainCharacteristics(long attack, long physicalDefense, long magicDefense) {
+    public MainCharacteristics(int attack, int physicalDefense, int magicDefense) {
         this.attack = attack;
         this.physicalDefense = physicalDefense;
         this.magicDefense = magicDefense;
     }
 
-    protected MainCharacteristics addAttack(long attack) {
+    protected MainCharacteristics addAttack(int attack) {
         this.attack += attack;
         return this;
     }
 
-    protected MainCharacteristics addPhysicalDefense(long physicalDefense) {
+    protected MainCharacteristics addPhysicalDefense(int physicalDefense) {
         this.physicalDefense += physicalDefense;
         return this;
     }
 
-    protected MainCharacteristics addMagicDefense(long magicDefense) {
+    protected MainCharacteristics addMagicDefense(int magicDefense) {
         this.magicDefense += magicDefense;
         return this;
     }
 
-    protected MainCharacteristics addSpellPower(long spellPower) {
+    protected MainCharacteristics addSpellPower(int spellPower) {
         this.spellPower += spellPower;
         return this;
     }
 
-    protected MainCharacteristics addMana(long mana) {
+    protected MainCharacteristics addMana(int mana) {
         this.mana += mana;
         return this;
     }
 
-    public long getAttack() {
+    public int getAttack() {
         return attack;
     }
 
-    public long getPhysicalDefense() {
+    public int getPhysicalDefense() {
         return physicalDefense;
     }
 
-    public long getMagicDefense() {
+    public int getMagicDefense() {
         return magicDefense;
     }
 
-    public long getSpellPower() {
+    public int getSpellPower() {
         return spellPower;
     }
 
-    public long getMana() {
+    public int getMana() {
         return mana;
     }
 }
